+++ packages/@runlightyear/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/cli
 
+## 0.11.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.12.3
+
 ## 0.11.0
 
 ### Minor Changes